+++ src/validation/nameValidator.ts
@@ -2,6 +2,10 @@ function hasBlank(str: string) {
   return str.trim() !== str;
 }
 
+function isEmpty(str: string) {
+  return str === '';
+}
+
 interface Validator {
   test: (str: string) => boolean;
   message: string;
@@ -13,10 +17,14 @@ const validators: Validator[] = [
     test: hasBlank,
     message: '이름 앞 뒤, 공백을 제외하고 입력해주세요.',
   },
+  {
+    test: isEmpty,
+    message: '이름을 입력해주세요.',
+  },
 ];
 
 // @TODO: never | void type 지정
-export function signinValidator(name: string) {
+export function nameValidator(name: string) {
   return validators.every(({ test, message }) => {
     if (test(name)) {
       throw new Error(message);
